requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-login==0.6.3
flask-wtf==1.2.1
python-dotenv==1.0.0
google-generativeai==0.3.2
opencv-python-headless==4.7.0.72
pillow==9.5.0
mediapipe==0.10.0
requests==2.28.2
werkzeug==3.0.1